0.9: Primera estructura, se modifica ID de la entidad Price

+++ src/main/java/inditex/productpriceservice/services/PricesService.java
@@ -15,7 +15,7 @@ public class PricesService {
     private PricesRepository pricesRepository;
 
     public Iterable<Price> findByBrandIdAndStartDateAndProductId(String brandId, Date startDate, String productId) {
-        int brandIdInt = Integer.parseInt(brandId);
+        long brandIdInt = Long.parseLong(brandId);
         return pricesRepository.findByBrandIdAndStartDateAndProductId(brandIdInt, startDate, productId);
     }
 }
